renamed ssa_conversion file
